FINALLY fix the vexing bug that made only queryparams count when calcing the webhookId
OK, so this actually kinda makes me mad. Express req objects
extend the http IncomingMessage object from nodejs. So,
I looked at the IncomingMessage docs, and found the req
property. Assuming that it would work as described in the
nodejs docs, I used it here. Dumb of me to assume that, I guess.
Turns out, that express screws with that property, and adds another
param (originalUrl) to actually refer to the URL. WHY? I don't know.
Honestly, why mess with the original spec when extending it? why add
an "originalUrl" field, and not a "mangledFuckedUpUrl" field with the
modified URL instead? A little wierdchamp, if I'm being honest. But
I found it out, don't worry. And yes, the Express docs do mention this
under their description for the originalUrl property.